Price bug fixed

+++ agency/Bike.java
@@ -1,14 +1,14 @@
 package agency;
 
 public class Bike extends Vehicle{
-    private static int _price = 30;
+    private static double _price = 30;
     private static int _maxtime = 15;
 
     public Bike(int id, String brand, String model, int year, boolean available){
         super(id, brand, model, year, available);
     }
 
-    public int getPrice(){
+    public double getPrice(){
         return _price;
     }
